+++ source/packages/@aws-accelerator/accelerator/lib/stacks/network-stacks/network-prep-stack/resolver-resources.ts
@@ -221,17 +221,17 @@ export class ResolverResources {
   ): Map<string, string> {
     const queryLogsMap = new Map<string, string>();
 
+    const centralLogsBucket = cdk.aws_s3.Bucket.fromBucketName(
+      this.stack,
+      'CentralLogsBucket',
+      `${
+        this.stack.acceleratorResourceNames.bucketPrefixes.centralLogs
+      }-${props.accountsConfig.getLogArchiveAccountId()}-${props.centralizedLoggingRegion}`,
+    );
+
     if (logItem && delegatedAdminAccountId === cdk.Stack.of(this.stack).account) {
       if (logItem.destinations.includes('s3')) {
         this.stack.addLogs(LogLevel.INFO, `Create DNS query log ${logItem.name}-s3 for central S3 destination`);
-        const centralLogsBucket = cdk.aws_s3.Bucket.fromBucketName(
-          this.stack,
-          'CentralLogsBucket',
-          `${
-            this.stack.acceleratorResourceNames.bucketPrefixes.centralLogs
-          }-${props.accountsConfig.getLogArchiveAccountId()}-${props.centralizedLoggingRegion}`,
-        );
-
         const s3QueryLogConfig = this.createQueryLogItem(logItem, centralLogsBucket, props.partition, orgId);
         queryLogsMap.set(`${logItem.name}-s3`, s3QueryLogConfig.logId);
       }
@@ -249,6 +249,40 @@ export class ResolverResources {
 
         const cwlQueryLogConfig = this.createQueryLogItem(logItem, logGroup, props.partition, orgId);
         queryLogsMap.set(`${logItem.name}-cwl`, cwlQueryLogConfig.logId);
+      }
+    }
+    for (const vpcItem of props.networkConfig.vpcs ?? []) {
+      const queryLogName = vpcItem.vpcRoute53Resolver?.queryLogs?.name;
+      if (vpcItem.vpcRoute53Resolver?.queryLogs) {
+        if (vpcItem.vpcRoute53Resolver.queryLogs.destinations.includes('s3')) {
+          this.stack.addLogs(LogLevel.INFO, `Create DNS query log ${queryLogName}-s3 for central S3 destination`);
+          const s3QueryLogConfig = this.createQueryLogItem(
+            vpcItem.vpcRoute53Resolver.queryLogs,
+            centralLogsBucket,
+            props.partition,
+            orgId,
+          );
+          queryLogsMap.set(`${queryLogName}-s3`, s3QueryLogConfig.logId);
+        }
+      }
+      if (vpcItem.vpcRoute53Resolver?.queryLogs?.destinations.includes('cloud-watch-logs')) {
+        this.stack.addLogs(
+          LogLevel.INFO,
+          `Create DNS query log ${queryLogName}-cwl for central CloudWatch logs destination`,
+        );
+
+        const logGroup = new cdk.aws_logs.LogGroup(this.stack, pascalCase(`${queryLogName}QueryLogsLogGroup`), {
+          encryptionKey: this.stack.cloudwatchKey,
+          retention: this.stack.logRetention,
+        });
+
+        const cwlQueryLogConfig = this.createQueryLogItem(
+          vpcItem.vpcRoute53Resolver?.queryLogs,
+          logGroup,
+          props.partition,
+          orgId,
+        );
+        queryLogsMap.set(`${queryLogName}-cwl`, cwlQueryLogConfig.logId);
       }
     }
     return queryLogsMap;